+++ catalystwan/endpoints/configuration/policy/list/color.py
@@ -5,17 +5,9 @@ from uuid import UUID
 
 from catalystwan.endpoints import APIEndpoints, delete, get, post, put
 from catalystwan.endpoints.configuration.policy.abstractions import PolicyListEndpoints
-from catalystwan.models.policy.lists import ColorList
-from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListInfo, PolicyListPreview
+from catalystwan.models.policy.list.color import ColorList, ColorListEditPayload, ColorListInfo
+from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListPreview
 from catalystwan.typed_list import DataSequence
-
-
-class ColorListEditPayload(ColorList, PolicyListId):
-    pass
-
-
-class ColorListInfo(ColorList, PolicyListInfo):
-    pass
 
 
 class ConfigurationPolicyColorList(APIEndpoints, PolicyListEndpoints):
