+++ catalystwan/endpoints/configuration/policy/list/community.py
@@ -5,17 +5,9 @@ from uuid import UUID
 
 from catalystwan.endpoints import APIEndpoints, delete, get, post, put
 from catalystwan.endpoints.configuration.policy.abstractions import PolicyListEndpoints
-from catalystwan.models.policy.lists import CommunityList
-from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListInfo, PolicyListPreview
+from catalystwan.models.policy.list.communities import CommunityList, CommunityListEditPayload, CommunityListInfo
+from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListPreview
 from catalystwan.typed_list import DataSequence
-
-
-class CommunityListEditPayload(CommunityList, PolicyListId):
-    pass
-
-
-class CommunityListInfo(CommunityList, PolicyListInfo):
-    pass
 
 
 class ConfigurationPolicyCommunityList(APIEndpoints, PolicyListEndpoints):
