+++ catalystwan/endpoints/configuration/policy/list/expanded_community.py
@@ -5,17 +5,13 @@ from uuid import UUID
 
 from catalystwan.endpoints import APIEndpoints, delete, get, post, put
 from catalystwan.endpoints.configuration.policy.abstractions import PolicyListEndpoints
-from catalystwan.models.policy.lists import ExpandedCommunityList
-from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListInfo, PolicyListPreview
+from catalystwan.models.policy.list.communities import (
+    ExpandedCommunityList,
+    ExpandedCommunityListEditPayload,
+    ExpandedCommunityListInfo,
+)
+from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListPreview
 from catalystwan.typed_list import DataSequence
-
-
-class ExpandedCommunityListEditPayload(ExpandedCommunityList, PolicyListId):
-    pass
-
-
-class ExpandedCommunityListInfo(ExpandedCommunityList, PolicyListInfo):
-    pass
 
 
 class ConfigurationPolicyExpandedCommunityList(APIEndpoints, PolicyListEndpoints):
